add validation and make some small improvements

+++ CodingTracker.KacperZys/CodingTracker.KacperZys/Controllers/MainMenuController.cs
@@ -1,51 +1,56 @@
-﻿using CodingTracker.KacperZys.Model;
+﻿using CodingTracker.KacperZys.Exceptions;
+using CodingTracker.KacperZys.Model;
 using Spectre.Console;
 
 namespace CodingTracker.KacperZys.Controllers;
 internal class MainMenuController
 {
     MainMenuModel mainMenuModel = new MainMenuModel();
+    CodingSessionCotroller CodingSessionCotroller = new();
     public void ViewAll()
     {
         List<CodingSession> sessions = mainMenuModel.ViewAll();
 
         foreach (var session in sessions)
         {
-            AnsiConsole.MarkupLine($"[yellow]ID: {session.Id}, Start Time: {session.StartTime}, End Time: {session.EndTime}, Duration: {session.Duration}[/]");
+            AnsiConsole.MarkupLine($"[yellow]ID: {session.Id}, Start Time: {session.StartTime.ToString("yyyy:MM:dd")}, End Time: {session.EndTime.ToString("yyyy:MM:dd")}" +
+                $", Duration: {session.Duration} hours[/]");
         }
     }
 
     public void CreateSession()
     {
-        mainMenuModel.Create(AskForSessionParams());
+        while (true)
+        {
+            try
+            {
+                mainMenuModel.Create(CodingSessionCotroller.AskForSessionParams());
+                break;
+            }
+            catch (WrongDateException ex)
+            {
+                AnsiConsole.MarkupLine($"[red]{ex.Message}[/]");
+            }
+        }
+
+
         AnsiConsole.Clear();
         AnsiConsole.MarkupLine("[green]Session added successfully.[/]");
-    }
-
-    private CodingSession AskForSessionParams()
-    {
-        CodingSession newSession = new CodingSession();
-        newSession.StartTime = DateTime.Parse(AnsiConsole.Ask<string>("Enter starting date: "));
-        newSession.EndTime = DateTime.Parse(AnsiConsole.Ask<string>("Enter ending date: "));
-        newSession.Duration = (newSession.EndTime - newSession.StartTime).ToString();
-
-        return newSession;
     }
 
     public void Delete()
     {
         int id = AnsiConsole.Ask<int>("Enter the ID of the session to delete: ");
-        bool deletionStatus = mainMenuModel.Delete(new SessionRequest { Id = id });
-        AnsiConsole.Clear();
 
-        if (deletionStatus)
-        {
-            AnsiConsole.MarkupLine("[green]Session deleted successfully.[/]");
-        }
-        else
+        if (mainMenuModel.Exists(new SessionRequest { Id = id }) == false)
         {
             AnsiConsole.MarkupLine("[red]No session found with the provided ID.[/]");
+            return;
         }
+
+        mainMenuModel.Delete(new SessionRequest { Id = id });
+        AnsiConsole.Clear();
+        AnsiConsole.MarkupLine("[green]Session deleted successfully.[/]");
     }
 
     public void Modify()
@@ -58,7 +63,7 @@ internal class MainMenuController
             return;
         }
 
-        CodingSession updatedSession = AskForSessionParams();
+        CodingSession updatedSession = CodingSessionCotroller.AskForSessionParams();
         updatedSession.Id = id;
 
         bool ModifyStatus = mainMenuModel.Modify(updatedSession);
